pyproject.toml
[tool.poetry]
name = "task-manager"
version = "0.1.0"
description = ""
authors = ["Freddy Morales <freddy_alexander@outlook.com>"]
readme = "README.md"
packages = [{include = "task_manager", from = "src"}]

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.4"
djangorestframework = "^3.15.2"
pymongo = "^4.10.1"
drf-yasg = "^1.21.8"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-django = "^4.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
